ft_isalpha ok

+++ libft/ft_strlen.c
@@ -6,29 +6,28 @@
 /*   By: aaugu <marvin@42lausanne.ch>               +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/24 12:03:19 by aaugu             #+#    #+#             */
-/*   Updated: 2022/10/25 14:08:55 by aaugu            ###   ########.fr       */
+/*   Updated: 2022/10/26 12:28:39 by aaugu            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-/*
+#include <stddef.h>
 #include <stdio.h>
-*/
-int	ft_strlen(const char *s)
+
+size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
 		i++;
 	return (i);
 }
-/*
+
 int	main(int argc, char **argv)
 {
 	if (argc != 2)
 		printf("Error\n");
 	else
-		printf("%d\n", ft_strlen(argv[1]));
+		printf("%ld\n", ft_strlen(argv[1]));
 	return (0);
 }
-*/
